# ecotrak-backend/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecotrak API",
    "description": "API for calculating GHG emissions",
    "version": "1.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/elecdata": {
      "post": {
        "tags": [
          "electrical"
        ],
        "summary": "calculate and add electricity data",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "state": {
                  "type": "string"
                },
                "elec": {
                  "type": "number"
                },
                "unit": {
                  "type": "string"
                }
              },
              "required": [
                "state",
                "elec",
                "unit"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The added electricity data",
            "schema": {
              "$ref": "#/definitions/ElecData"
            }
          }
        }
      }
    },
    "/sc2data/{state}/{unit}": {
      "get": {
        "tags": [
          "electrical"
        ],
        "summary": "Get a list of scope 2 data for a particular state and unit",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "unit",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of scope 2 data",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        }
      }
    },
    "/sc3data/{state}/{unit}": {
      "get": {
        "tags": [
          "electrical"
        ],
        "summary": "Get a list of scope 3 data for a particular state and unit",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "unit",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of scope 3 data",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        }
      }
    },
...
      "/fueldata": {
        "post": {
          "tags": ["fuel"],
          "summary": "Add fuel data to the database",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "fuel": {
                    "type": "number"
                  },
                  "fueltype": {
                    "type": "string"
                  },
                  "fuelsubtype": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  }
                },
                "required": [
                  "fuel",
                  "fueltype",
                  "fuelsubtype",
                  "unit"
                ]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "type": "object",
                "properties": {
                  "fuel": {
                    "type": "string"
                  },
                  "fueltype": {
                    "type": "string"
                  },
                  "fuelsubtype": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  },
                  "total": {
                    "type": "number",
                    "format": "float"
                  },
                  "CO2": {
                    "type": "number",
                    "format": "float"
                  },
                  "CH4": {
                    "type": "number",
                    "format": "float"
                  },
                  "N2O": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            }
          }
          }
        },
...
      "/wastedata": {
        "post": {
          "tags": [
            "waste"
          ],
          "summary": "Add waste data to the database",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "waste": {
                    "type": "number"
                  },
                  "type": {
                    "type": "string"
                  },
                  "subtype": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  }
                },
                "required": [
                  "waste",
                  "type",
                  "subtype",
                  "unit"
                ]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "type": "object",
                "properties": {
                  "waste": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  },
                  "result": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            }
          }
        }
      },
...
      "/transdata": {
        "post": {
          "tags": [
            "transport"
          ],
          "summary": "calculate and add transportation data",
          "parameters": [
            {
              "name": "data",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "trans": {
                    "type": "number"
                  },
                  "transtype": {
                    "type": "string"
                  },
                  "fueltype": {
                    "type": "string"
                  },
                  "unit": {
                    "type": "string"
                  }
                },
                "required": [
                  "trans",
                  "transtype",
                  "fueltype",
                  "unit"
                ]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully added transportation data",
              "schema": {
                "$ref": "#/definitions/TransData"
              }
            }
          }
        }
      },
...
      "/uselecdata": {
        "post": {
          "tags": [
            "US electrical"
          ],
          "summary": "Calculate and add electricity usage data",
          "parameters": [
            {
              "name": "data",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "region": {
                    "type": "string"
                  },
                  "elec": {
                    "type": "number"
                  }
                },
                "required": [
                  "region",
                  "elec"
                ]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Electricity usage data added successfully",
              "schema": {
                "type": "object",
                "properties": {
                  "region": {
                    "type": "string"
                  },
                  "CO2": {
                    "type": "number"
                  },
                  "CH4": {
                    "type": "number"
                  },
                  "N2O": {
                    "type": "number"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input parameters"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "ElecData": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "state": {"type": "string"},
          "elec": {"type": "number"},
          "unit": {"type": "string"},
          "result": {"type": "number"}
        }
      },
...
      "TransData": {
        "type": "object",
        "properties": {
          "fueltype": {
            "type": "string"
          },
          "transtype": {
            "type": "string"
          },
          "trans": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "CO2": {
            "type": "number"
          },
          "CH4": {
            "type": "number"
          },
          "N2O": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        }
        },
...
      }
  }